refactoring into packages, added run-them-all root script
