+++ elastic/retrieve_wikipedia.py
@@ -1,0 +1,11 @@
+"""
+Provides helper functions for retrieving data from ES indices of wikipedia articles.
+"""
+
+
+def get_all_text(es_index_name):
+    pass
+
+
+def get_all_opening_text(es_index_name):
+    pass
